feat(inspector): restructure performance page components and add tests
