+++ packages/jazz-tools/src/inspector/pages/performance/index.ts
@@ -1,0 +1,2 @@
+export { PerformancePage } from "./PerformancePage.js";
+export type { PerformancePageProps } from "./PerformancePage.js";
