+++ packages/jazz-tools/src/inspector/pages/performance/types.ts
@@ -1,0 +1,12 @@
+export interface SubscriptionEntry {
+  uuid: string;
+  id: string;
+  source: string;
+  resolve: string;
+  status: "pending" | "loaded" | "error";
+  startTime: number;
+  endTime?: number;
+  duration?: number;
+  errorType?: string;
+  callerStack?: string;
+}
